+++ R/na_ma.R
@@ -133,7 +133,7 @@ na_ma <- function(x, k = 4, weighting = "exponential", maxgap = Inf) {
 
     # 1.3 Check for algorithm specific minimum amount of non-NA values
     if (sum(!missindx) < 2) {
-      stop("Input data needs at least 2 non-NA data point for applying na_ma")
+      stop("Input data needs at least 2 non-NA data points for applying na_ma")
     }
 
     # 1.4 Checks and corrections for wrong data dimension