+++ R/na_seasplit.R
@@ -108,7 +108,7 @@ na_seasplit <- function(x, algorithm = "interpolation", find_frequency = FALSE, 
 
     # 1.3 Check for algorithm specific minimum amount of non-NA values
     if (sum(!missindx) < 3) {
-      stop("Input data needs at least 3 non-NA data point for applying na_seasplit")
+      stop("Input data needs at least 3 non-NA data points for applying na_seasplit")
     }
 
     # 1.4 Checks and corrections for wrong data dimension